requirements.txt
anyio==3.6.1
appdirs==1.4.4
APScheduler==3.9.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
flake8==4.0.1
flake8-broken-line==0.4.0
flake8-isort==4.1.2.post0
flake8-plugin-utils==1.3.2
flake8-return==1.1.3
greenlet==1.1.2
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
importlib-metadata==4.12.0
isort==5.10.1
mccabe==0.6.1
pep8-naming==0.13.1
psycopg-binary==3.0.16
pycodestyle==2.8.0
pyee==8.2.2
pyflakes==2.4.0
pyppeteer==1.0.2
python-dotenv==0.20.0
python-telegram-bot==20.0a2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
tornado==6.2
tqdm==4.64.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.11
websockets==10.3
zipp==3.8.1
